security: Require admin authentication to load the consent statistics page.
The statistics page could potentially take a long time to load if the amount of records in the backend is huge. An adequate number of requests to this page could therefore allow a denial of service of the entire IdP if the attacker manages to send as many requests as processes available.

This resolves #1.